requirements.txt
# requirements.txt
psycopg2-binary==2.9.11
pandas==2.3.3
openpyxl==3.1.5
PyQt5==5.15.11
python-dotenv==1.2.1